Move isRelativePath to Util and use from hudson.model.Slave's descriptor
